pyproject.toml
[project]
name = "landscaping-app"
version = "0.1.0"
description = "Django scheduling application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "psycopg[binary]>=3.2.12",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "ruff>=0.14.2",
]
